feat: refactor user service imports and rename result files for improved structure
