+++ shared/src/database/utils/result.ts
